requirements.txt
pandas
numpy
matplotlib
seaborn
scikit-learn
jupyter
python-dotenv
psycopg2-binary

